server/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database (direct PostgreSQL connection)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Validation and settings
pydantic>=2.5.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Testing
pytest>=7.0.0
httpx>=0.25.0

# ===== PaperRefinery - PDF Processing Pipeline =====

# PDF Parsing
pymupdf>=1.24.0

# LLM - Together.ai (Llama models)
together>=1.0.0

# Embeddings - OpenAI
openai>=1.0.0

# Database - Supabase Python client
supabase>=2.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0
